requirements.txt
blinker==1.6.2
click==8.1.6
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Login==0.6.2
Flask-WTF==1.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pymongo==4.4.1
python-dotenv==1.0.0
Werkzeug==2.3.7
WTForms==3.0.1
